Pole

+++ Formula-Bollo-App/src/main/java/formula/bollo/app/controller/PenaltyController.java
@@ -3,11 +3,15 @@ package formula.bollo.app.controller;
 import formula.bollo.app.entity.Driver;
 import formula.bollo.app.entity.Penalty;
 import formula.bollo.app.entity.Race;
+import formula.bollo.app.entity.Season;
 import formula.bollo.app.mapper.PenaltyMapper;
+import formula.bollo.app.mapper.SeasonMapper;
 import formula.bollo.app.model.DriverPenaltiesDTO;
 import formula.bollo.app.model.PenaltyDTO;
+import formula.bollo.app.model.SeasonDTO;
 import formula.bollo.app.repository.PenaltyRepository;
 import formula.bollo.app.repository.RaceRepository;
+import formula.bollo.app.repository.SeasonRepository;
 import formula.bollo.app.services.PenaltyService;
 import formula.bollo.app.utils.Constants;
 import formula.bollo.app.utils.Log;
@@ -33,7 +37,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-@CrossOrigin(origins = Constants.URL_FRONTED)
+@CrossOrigin(origins = Constants.PRODUCTION_FRONTEND)
 @RestController
 @RequestMapping(path = {Constants.ENDPOINT_PENALTY}, produces = MediaType.APPLICATION_JSON_VALUE)
 @Tag(name = Constants.TAG_PENALTY, description = Constants.TAG_PENALTY_SUMMARY)
@@ -51,12 +55,20 @@ public class PenaltyController {
     @Autowired
     private PenaltyService penaltyService;
 
+    @Autowired
+    private SeasonRepository seasonRepository;
+
+    @Autowired
+    private SeasonMapper seasonMapper;
+
     @Operation(summary = "Get all penalties", tags = Constants.TAG_PENALTY)
     @GetMapping("/all")
-    public List<PenaltyDTO> getAllPenalties() {
+    public List<PenaltyDTO> getAllPenalties(@RequestParam(value = "season", required = false) Integer season) {
         Log.info("START - getAllPenalties - START");
-        
-        List<Penalty> penalties = penaltyRepository.findAll();
+        Log.info("RequestParam getAllPenalties (season) -> " + season);
+
+        int numberSeason = season == null ? Constants.ACTUAL_SEASON : season;
+        List<Penalty> penalties = penaltyRepository.findBySeason(numberSeason);
         List<PenaltyDTO> penaltyDTOs = new ArrayList<>();
 
         if (penalties.isEmpty()) return penaltyDTOs;
@@ -69,15 +81,17 @@ public class PenaltyController {
 
     @Operation(summary = "Get penalties per circuit", tags = Constants.TAG_PENALTY)
     @GetMapping("/circuit")
-    public List<PenaltyDTO> getPenaltiesByCircuit(@RequestParam("circuitId") Integer circuitId) {
+    public List<PenaltyDTO> getPenaltiesByCircuit(@RequestParam("circuitId") Integer circuitId, @RequestParam(value = "season", required = false) Integer season) {
         Log.info("START - getPenaltiesByCircuit - START");
         Log.info("RequestParam getPenaltiesByCircuit (circuitId) -> " + circuitId);
+        Log.info("RequestParam getPenaltiesByCircuit (season) -> " + season);
 
-        List<Race> races = raceRepository.findByCircuitId((long) circuitId);
+        int numberSeason = season == null ? Constants.ACTUAL_SEASON : season;
+        List<Race> races = raceRepository.findByCircuitId((long) circuitId, numberSeason);
         List<PenaltyDTO> penaltyDTOs = new ArrayList<>();
 
         if (races.isEmpty()) return penaltyDTOs;
-        List<Penalty> penalties = penaltyRepository.findByRaceId(races.get(0).getId());
+        List<Penalty> penalties = penaltyRepository.findByRaceId(races.get(0).getId(), numberSeason);
 
         if (penalties.isEmpty()) return penaltyDTOs;
         penaltyDTOs = penaltyMapper.convertPenaltiesToPenaltiesDTO(penalties);
@@ -89,11 +103,14 @@ public class PenaltyController {
 
     @Operation(summary = "Get penalties per driver", tags = Constants.TAG_PENALTY)
     @GetMapping("/totalPerDriver")
-    public List<DriverPenaltiesDTO> getPenaltiesByDriverAndRace() {
+    public List<DriverPenaltiesDTO> getPenaltiesByDriverAndRace(@RequestParam(value = "season", required = false) Integer season) {
         Log.info("START - getPenaltiesByDriverAndRace - START");
-    
+        Log.info("RequestParam getPenaltiesByDriverAndRace (season) " + season);
+        
+        int numberSeason = season == null ? Constants.ACTUAL_SEASON : season;
+
         List<DriverPenaltiesDTO> driversWithPenalties = new ArrayList<>();
-        List<Penalty> penalties = penaltyRepository.findAll();
+        List<Penalty> penalties = penaltyRepository.findBySeason(numberSeason);
 
         if (penalties.isEmpty()) return driversWithPenalties;
         Map<Driver, Map<Race, List<Penalty>>> driverRacePenaltiesMap = new HashMap<>();
@@ -108,13 +125,21 @@ public class PenaltyController {
 
     @Operation(summary = "Save penalties", tags = Constants.TAG_PENALTY)
     @PutMapping(path = "/save", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> savePenalty(@RequestBody List<PenaltyDTO> penaltyDTOs) {
+    public ResponseEntity<String> savePenalty(@RequestBody List<PenaltyDTO> penaltyDTOs, @RequestParam(value = "season", required = false) Integer season) {
         Log.info("START - savePenalty - START");        
         Log.info("RequestBody savePenalty " + penaltyDTOs.toString());
+        Log.info("RequestParam savePenalty (season) " + season);
 
         try {
+            int numberSeason = season == null ? Constants.ACTUAL_SEASON : season;
             Penalty firstPenalty = penaltyMapper.penaltyDTOToPenalty(penaltyDTOs.get(0));
-            List<Penalty> existingPenalties = penaltyRepository.findByDriverAndRaceAndSeverity(firstPenalty.getDriver().getId(), firstPenalty.getRace().getId(), firstPenalty.getSeverity().getId());
+            List<Penalty> existingPenalties = penaltyRepository.findByDriverAndRaceAndSeverity(firstPenalty.getDriver().getId(), firstPenalty.getRace().getId(), firstPenalty.getSeverity().getId(), numberSeason);
+           
+            List<Season> seasons = this.seasonRepository.findByNumber(numberSeason);
+            if (seasons.isEmpty()) return new ResponseEntity<>(Constants.ERROR_SEASON, Constants.HEADERS_TEXT_PLAIN, HttpStatusCode.valueOf(500));
+            SeasonDTO seasonToSave = this.seasonMapper.seasonToSeasonDTO(seasons.get(0));
+            penaltyDTOs.forEach((PenaltyDTO penaltyDTO) -> penaltyDTO.setSeason(seasonToSave));
+
             penaltyRepository.deleteAll(existingPenalties);
             penaltyService.savePenalties(penaltyDTOs);
         } catch (DataAccessException e) {
@@ -132,14 +157,16 @@ public class PenaltyController {
 
     @Operation(summary = "Get a penalty for a driver and circuit", tags = Constants.TAG_PENALTY)
     @GetMapping("/perDriverPerRace")
-    public List<PenaltyDTO> getPenaltyByDriverAndRace(@RequestParam("driverId") Integer driverId, @RequestParam("raceId") Integer raceId, @RequestParam("severityId") Integer severityId) {
+    public List<PenaltyDTO> getPenaltyByDriverAndRace(@RequestParam("driverId") Integer driverId, @RequestParam("raceId") Integer raceId, @RequestParam("severityId") Integer severityId, @RequestParam(value = "season", required = false) Integer season) {
         Log.info("START - getPenaltyByDriverAndRace - START");
         Log.info("RequestParam getPenaltyByDriverAndRace (driverId) -> " + driverId);
         Log.info("RequestParam getPenaltyByDriverAndRace (raceId) -> " + raceId);
         Log.info("RequestParam getPenaltyByDriverAndRace (severityId) -> " + severityId);
+        Log.info("RequestParam getPenaltyByDriverAndRace (season) -> " + season);
 
+        int numberSeason = season == null ? Constants.ACTUAL_SEASON : season;
         List<PenaltyDTO> penaltyDTOs = new ArrayList<>();
-        List<Penalty> penaltiesDriver = penaltyRepository.findByDriverAndRaceAndSeverity((long) driverId, (long) raceId, (long) severityId);
+        List<Penalty> penaltiesDriver = penaltyRepository.findByDriverAndRaceAndSeverity((long) driverId, (long) raceId, (long) severityId, numberSeason);
 
         if (penaltiesDriver.isEmpty()) return penaltyDTOs;
         penaltyDTOs = penaltyMapper.convertPenaltiesToPenaltiesDTO(penaltiesDriver);
